Adding "unsigned int" type casting to comparison
Due to a compilation flag on the Linux build, any comparison must have
been done with the same type ([-Werror=sign-compare]).

+++ cocos2dx/CCDirector.cpp
@@ -595,11 +595,11 @@ void CCDirector::popToSceneStackLevel(int level)
     }
 
     // current level or lower -> nothing
-    if (level >= c)
+    if ((unsigned int)level >= c)
         return;
 
 	// pop stack until reaching desired level
-	while (c > level)
+	while (c > (unsigned int)level)
     {
 		CCScene *current = (CCScene*)m_pobScenesStack->lastObject();
